Re-add public `!guide` command in party chat
